requirements.txt
streamlit==1.50.0
google-generativeai==0.8.5
python-dotenv==1.2.1
PyPDF2==3.0.1